% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auditcheck}
\alias{auditcheck}
\title{perform checks using audit files}
\usage{
auditcheck(pathtofolder, excluded = NULL)
}
\arguments{
\item{pathtofolder}{path to all audit csv files folder}

\item{excluded}{variables you want to exclude when calculating survey duration}
}
\value{
Data frame with survey duration and other useful information
}
\description{
perform checks using audit files
}
