% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_survey_duration}
\alias{calculate_survey_duration}
\title{Calculate survey duration}
\usage{
calculate_survey_duration(endtime, startime)
}
\arguments{
\item{endtime}{Column end}

\item{startime}{column start}
}
\value{
Duration in minutes
}
\description{
Calculate survey duration
}
