% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_event_type}
\alias{count_event_type}
\title{Count occurrence of a specified type}
\usage{
count_event_type(allaudit, event)
}
\arguments{
\item{allaudit}{List of all audit files}

\item{event}{a specific event}
}
\value{
Number of specified event if happened
}
\description{
Count occurrence of a specified type
}
