% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{form_lifetime}
\alias{form_lifetime}
\title{Calculates form lifetime}
\usage{
form_lifetime(endtime, startime)
}
\arguments{
\item{endtime}{column end}

\item{startime}{column start}
}
\value{
Duration in minutes
}
\description{
Calculates form lifetime
}
