% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_all_csv}
\alias{read_all_csv}
\title{Read all csv files in a folder
\code{read_all_csv} returns a list of all csv files located inside folder in its arguments}
\usage{
read_all_csv(folder)
}
\arguments{
\item{folder}{A character string. Path to folder containing csv files}
}
\value{
The output is a list of all CSV files inside the argument folder
}
\description{
Read all csv files in a folder
\code{read_all_csv} returns a list of all csv files located inside folder in its arguments
}
