% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_node}
\alias{sanitize_node}
\title{Delete all characters before the last slash including the slash}
\usage{
sanitize_node(expression)
}
\arguments{
\item{expression}{column of a dataframe}
}
\description{
Delete all characters before the last slash including the slash
}
